Refactored config and tests for config and validator
